petals/requirements.txt
# Core dependencies
sanic==23.6.0
torch==2.1.0
transformers==4.34.1
diffusers==0.21.4
ipfshttpclient==0.8.0
aiofiles==23.2.1
Pillow==10.0.1
petals-client==2.2.0
web3==6.11.1
eth-account==0.9.0

# Cloud deployment dependencies
boto3>=1.28.0  # AWS
google-cloud-compute>=1.10.0  # GCP
azure-mgmt-compute>=29.1.0  # Azure
azure-mgmt-resource>=23.0.0
azure-identity>=1.13.0
